fix: core when co_await move assigned awaiters
In the previous implementation, the `this` pointer of the awaiter was
captured into the callback function of workflow's task. When the awaiter
was moved assigned to another object, the `this` captured into the callback
function would be destroyed. Calling the callback function at this time
would cause problems.

Added BasicAwaiter and AwaiterInfo as a bridge, reacquire the current
awaiter in the callback function to avoid using destroyed objects.